requirements-test.txt
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
factory-boy>=3.3.0
Faker>=18.0.0
httpx>=0.28.0
sqlalchemy-utils>=0.41.0
fastapi>=0.115.0
starlette>=0.41.0